# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Translation Management Service API",
        "description": "API for managing translations with multi-language support",
        "version": "1.0.0"
    },
    "paths": {
        "/api/translations/{id}": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Get a translation by ID",
                "operationId": "f288295c6f30e957c7769d28752f7905",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Translation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Translation"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Translation not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Translations"
                ],
                "summary": "Update a translation",
                "description": "Updates an existing translation record",
                "operationId": "updateTranslation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the translation to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Translation data to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "group",
                                    "key",
                                    "value",
                                    "locale"
                                ],
                                "properties": {
                                    "group": {
                                        "type": "string",
                                        "example": "validation"
                                    },
                                    "key": {
                                        "type": "string",
                                        "example": "required"
                                    },
                                    "value": {
                                        "type": "string",
                                        "example": "This field is required"
                                    },
                                    "locale": {
                                        "type": "string",
                                        "example": "en"
                                    },
                                    "tag": {
                                        "type": "string",
                                        "example": "web",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Translation updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Translation"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid"
                                        },
                                        "errors": {
                                            "properties": {
                                                "field_name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The field name is required"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Translation not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Translation not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Translations"
                ],
                "summary": "Deletes an existing translation by ID",
                "operationId": "76d12830146dc4acbde69c403b3d95e8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Translation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Translation"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Translation not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/translations": {
            "post": {
                "tags": [
                    "Translations"
                ],
                "summary": "Create a new translation",
                "description": "Stores a new translation record in the database",
                "operationId": "createTranslation",
                "requestBody": {
                    "description": "Translation data to store",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "group",
                                    "key",
                                    "value",
                                    "locale"
                                ],
                                "properties": {
                                    "group": {
                                        "description": "The translation group/category",
                                        "type": "string",
                                        "example": "validation"
                                    },
                                    "key": {
                                        "description": "The translation key",
                                        "type": "string",
                                        "example": "required"
                                    },
                                    "value": {
                                        "description": "The translated text",
                                        "type": "string",
                                        "example": "This field is required"
                                    },
                                    "locale": {
                                        "description": "Language code (2 characters)",
                                        "type": "string",
                                        "example": "en"
                                    },
                                    "tag": {
                                        "description": "Optional context tag",
                                        "type": "string",
                                        "example": "web",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Translation created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Translation"
                                },
                                "example": {
                                    "id": 1,
                                    "group": "validation",
                                    "key": "required",
                                    "value": "This field is required",
                                    "locale": "en",
                                    "tag": "web",
                                    "created_at": "2023-07-15T10:00:00.000000Z",
                                    "updated_at": "2023-07-15T10:00:00.000000Z"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "group": [
                                                    "The group field is required"
                                                ],
                                                "locale": [
                                                    "The locale must be 2 characters"
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict - Duplicate translation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Translation already exists for this group/key/locale combination"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/translations/search": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Search translations",
                "description": "Search translations by query, tag, or locale",
                "operationId": "searchTranslations",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Search term to look for in key or value",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tag",
                        "in": "query",
                        "description": "Filter by tag (web, mobile, desktop, tablet)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "locale",
                        "in": "query",
                        "description": "Filter by locale (en, fr, es)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "en"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Translation"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/translations/export": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Export translations",
                "description": "Export translations in JSON format grouped by locale and group",
                "operationId": "exportTranslations",
                "parameters": [
                    {
                        "name": "tag",
                        "in": "query",
                        "description": "Filter by tag (web, mobile, desktop, tablet)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "locale",
                        "in": "query",
                        "description": "Filter by specific locale",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "en"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "en": {
                                        "validation": {
                                            "required": "This field is required"
                                        }
                                    },
                                    "fr": {
                                        "validation": {
                                            "required": "Ce champ est obligatoire"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Translation": {
                "title": "Translation",
                "required": [
                    "group",
                    "key",
                    "value",
                    "locale"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "group": {
                        "type": "string",
                        "example": "validation"
                    },
                    "key": {
                        "type": "string",
                        "example": "required"
                    },
                    "value": {
                        "type": "string",
                        "example": "This field is required"
                    },
                    "locale": {
                        "type": "string",
                        "example": "en"
                    },
                    "tag": {
                        "type": "string",
                        "example": "web",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Translations",
            "description": "API Endpoints for managing translations"
        }
    ]
}